+++ core/server/api_container/server/service_network/service_network.go
@@ -223,6 +223,7 @@ func (network *ServiceNetwork) StartService(
 	serviceId service.ServiceID,
 	imageName string,
 	privatePorts map[string]*port_spec.PortSpec,
+	publicPorts map[string]*port_spec.PortSpec, //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
 	entrypointArgs []string,
 	cmdArgs []string,
 	dockerEnvVars map[string]string,
@@ -285,6 +286,7 @@ func (network *ServiceNetwork) StartService(
 		serviceGuid,
 		imageName,
 		privatePorts,
+		publicPorts,
 		entrypointArgs,
 		cmdArgs,
 		dockerEnvVars,
@@ -703,6 +705,7 @@ func (network *ServiceNetwork) startService(
 	serviceGuid service.ServiceGUID,
 	imageName string,
 	privatePorts map[string]*port_spec.PortSpec,
+	publicPorts map[string]*port_spec.PortSpec, //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
 	entrypointArgs []string,
 	cmdArgs []string,
 	envVars map[string]string,
@@ -789,6 +792,7 @@ func (network *ServiceNetwork) startService(
 		serviceGuid,
 		imageName,
 		privatePorts,
+		publicPorts,
 		entrypointArgs,
 		cmdArgs,
 		envVars,